requirements.txt
django==3.2.16
djangorestframework==3.14.0
drf-nested-routers==0.93.4
mysqlclient==2.0.1
django-mysql==3.8.1